Setting up data layer - get current weather
+ secret key
+ Hilt configuration  (Data module)
+ Api service
+ RepositoryImpl
+ mapper

+++ core/src/main/java/com/custom/core/model/WeatherModel.kt
@@ -6,7 +6,7 @@ data class WeatherModel(
     val temperatureModel: TemperatureModel,
     val conditionDescription: String,
     val humidityPercentage: Int,
-    val windSpeed: Int,
+    val windSpeed: Double,
     val iconUrl: String
 )
